+++ python/ray/tune/logger.py
@@ -13,7 +13,7 @@ from ray.tune.result import (NODE_IP, TRAINING_ITERATION, TIME_TOTAL_S,
                              EXPR_PARAM_PICKLE_FILE, EXPR_PROGRESS_FILE,
                              EXPR_RESULT_FILE)
 from ray.tune.syncer import get_node_syncer
-from ray.tune.utils import flatten_dict
+from ray.tune.utils import flatten_dict, log_once
 
 logger = logging.getLogger(__name__)
 
@@ -215,10 +215,11 @@ class TBXLogger(Logger):
                 # In case TensorboardX still doesn't think it's a valid value
                 # (e.g. `[[]]`), warn and move on.
                 except (ValueError, TypeError):
-                    logger.warning(
-                        "You are trying to log an invalid value ({}={}) "
-                        "via {}!".format(full_attr, value,
-                                         type(self).__name__))
+                    if log_once("invalid_tbx_value"):
+                        logger.warning(
+                            "You are trying to log an invalid value ({}={}) "
+                            "via {}!".format(full_attr, value,
+                                             type(self).__name__))
 
         self.last_result = valid_result
         self._file_writer.flush()
