+++ python/ray/tune/utils/__init__.py
@@ -1,9 +1,19 @@
-from ray.tune.utils.util import (deep_update, flatten_dict, get_pinned_object,
-                                 merge_dicts, pin_in_object_store, UtilMonitor,
-                                 validate_save_restore, warn_if_slow)
+from ray.tune.utils.debug import log_once, disable_log_once_globally, \
+    enable_periodic_logging
+from ray.tune.utils.util import deep_update, flatten_dict, get_pinned_object, \
+    merge_dicts, pin_in_object_store, UtilMonitor, validate_save_restore, \
+    warn_if_slow
 
 __all__ = [
-    "deep_update", "flatten_dict", "get_pinned_object", "merge_dicts",
-    "pin_in_object_store", "UtilMonitor", "validate_save_restore",
+    "deep_update",
+    "disable_log_once_globally",
+    "enable_periodic_logging",
+    "flatten_dict",
+    "get_pinned_object",
+    "log_once",
+    "merge_dicts",
+    "pin_in_object_store",
+    "UtilMonitor",
+    "validate_save_restore",
     "warn_if_slow"
 ]
