+++ rllib/policy/eager_tf_policy.py
@@ -13,8 +13,8 @@ from ray.rllib.policy.sample_batch import SampleBatch
 from ray.rllib.policy.policy import ACTION_PROB, ACTION_LOGP
 from ray.rllib.utils import add_mixins
 from ray.rllib.utils.annotations import override
-from ray.rllib.utils.debug import log_once
 from ray.rllib.utils import try_import_tf
+from ray.tune.utils.debug import log_once
 
 tf = try_import_tf()
 logger = logging.getLogger(__name__)
